// lib/types.ts
export type MeetingType =
    | 'testimony'
    | 'regular'
    | 'stake'
    | 'general';

export interface Hymn {
    number: number;
    title: string;
}

export interface Speaker {
    name: string;
    topic: string;
    type: 'speaker' | 'musical number';
}

export interface WardBusinessItem {
    description: string;
}

export interface Meeting {
    id: number;
    date: string;
    meetingType: MeetingType;
    presiding: string;
    conducting: string;
    announcements?: string[];
    openingHymn: Hymn;
    openingPrayer: string;
    wardBusiness: WardBusinessItem[];
    stakeBusiness: boolean;
    sacramentHymn: Hymn;
    speakers: Speaker[];
    closingHymn: Hymn;
    closingPrayer: string;
}